refactor : 장소상세 info 컴포넌트화 (#197)
